# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: StreamPro — Enterprise TV Streaming Platform API
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/phone + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout (revoke current session)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request OTP for password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
              required: [identifier]
      responses:
        "200":
          description: OTP dispatched (response never includes the code)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                code:
                  type: string
                type:
                  type: string
                  enum: [forgot_password, email_verification, phone_verification]
              required: [identifier, code, type]
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password after OTP verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                otpCode:
                  type: string
                newPassword:
                  type: string
              required: [identifier, otpCode, newPassword]
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileDto"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
              required: [currentPassword, newPassword]
      responses:
        "200":
          description: Password changed
  /auth/social:
    post:
      operationId: socialLogin
      tags: [auth]
      summary: Social login (Google / Apple / Facebook)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [google, apple, facebook]
                accessToken:
                  type: string
                email:
                  type: string
                name:
                  type: string
              required: [provider, accessToken]
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ── CHANNELS ─────────────────────────────────────────────────────────────
...
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Get channel by id
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update channel (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelDto"
      responses:
        "200":
          description: Updated channel
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete channel (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Deleted

  # ── EPG ──────────────────────────────────────────────────────────────────
  /epg:
    get:
      operationId: getEpg
      tags: [epg]
      summary: Get EPG entries
      parameters:
        - name: channelId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: EPG entries
...
  /advertisements/providers:
    get:
      operationId: getAdProviders
      tags: [advertisements]
      summary: List ad providers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider list
  /advertisements/placements:
    get:
      operationId: getAdPlacements
      tags: [advertisements]
      summary: List ad placements
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: device
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Placement list
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all settings
      responses:
        "200":
          description: Settings key-value map
    post:
      operationId: upsertSetting
      tags: [settings]
      summary: Create/update a setting (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
              required: [key, value]
      responses:
        "200":
          description: Upserted
  /settings/{key}:
    get:
      operationId: getSetting
      tags: [settings]
      summary: Get setting by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Setting value
    delete:
      operationId: deleteSetting
      tags: [settings]
      summary: Delete setting (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── PARENTAL CONTROL ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      required: [statusCode, message]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/TokenPair"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin, moderator]
        isActive:
          type: boolean
        isPremium:
          type: boolean
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        avatar:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        language:
          type: string
      required: [id, name, role, isActive, isPremium]
    UpdateProfileDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        language:
          type: string
        country:
          type: string
        avatar:
          type: string
    RegisterDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6
      required: [name, password]
    LoginDto:
      type: object
      properties:
        identifier:
          type: string
        password:
          type: string
        deviceName:
          type: string
        deviceType:
          type: string
        platform:
          type: string
      required: [identifier, password]
    Channel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        streamUrl:
          type: string
        logo:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required: [id, name, slug, streamUrl, isActive]
    CreateChannelDto:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        streamUrl:
          type: string
        logo:
          type: string
        country:
          type: string
        isActive:
          type: boolean
      required: [name, slug, streamUrl]
...